+++ spec/kitchen/driver/localhost_spec.rb
@@ -1,0 +1,11 @@
+# Encoding: UTF-8
+
+require_relative '../../spec_helper'
+require_relative '../../../lib/kitchen/driver/localhost'
+
+describe Kitchen::Driver::Localhost do
+  it 'does some things' do
+    pending
+    expect(true).to eq(false)
+  end
+end