+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.1.0 / Unreleased
+Kitchen-Localhost CHANGELOG
+===========================
 
-* Initial release
+v0.0.1 (2015-04-10)
+-------------------
+* Development started!